Split Search into SearchInputs & SearchResults
